requirements.txt
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.1.0
pandas==2.1.4
numpy==1.26.2
pyinstaller==6.3.0
requests==2.32.0
pyarrow==14.0.1